What type of PR is this?  for slot sharding rule What this PR does / why we need it:
Which issue(s) this PR fixes:

Special notes for your reviewer:

Does this PR introduce a user-facing change?:

+++ server/common/model/src/main/java/com/alipay/sofa/registry/common/model/slot/MD5SlotFunction.java
@@ -24,22 +24,22 @@ import com.alipay.sofa.registry.consistency.hash.MD5HashFunction;
  * @version v 0.1 2020-11-02 15:43 yuzhi.lyz Exp $
  */
 public final class MD5SlotFunction implements SlotFunction {
-    public final  short           MAX_SLOTS       = 16384;
-    private final short           maxSlots;
+    public final  int             MAX_SLOTS       = 16384;
+    private final int             maxSlots;
     private final MD5HashFunction md5HashFunction = new MD5HashFunction();
 
     public MD5SlotFunction() {
         this.maxSlots = MAX_SLOTS;
     }
 
-    @Override public short maxSlots() {
+    @Override public int maxSlots() {
         return maxSlots;
     }
 
     @Override
-    public short slotOf(Object o) {
+    public int slotOf(Object o) {
         // make sure >=0
         final int hash = Math.abs(md5HashFunction.hash(o));
-        return (short) (hash % maxSlots);
+        return hash % maxSlots;
     }
 }
